fix semicolons in extend.texi
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/ibm@241038 138bc75d-0d04-0410-961f-82ee72b054a4

+++ rfc2467.2/gcc/testsuite/gcc.target/powerpc/vsu/vec-xst-len-4.c
@@ -8,9 +8,10 @@
 #include <altivec.h>
 
 void
-store_data (vector unsigned int *datap, unsigned int *address, size_t length)
+store_data (vector signed __int128 *datap, signed __int128 *address,
+	    size_t length)
 {
-  vector unsigned int data = *datap;
+  vector signed __int128 data = *datap;
 
   vec_xst_len (data, address, length);
 }
